+++ api/database/migrations/2022_09_27_123402_create_vehicle_service_tokens_table.php
@@ -16,7 +16,7 @@ class CreateVehicleServiceTokensTable extends Migration
         Schema::create('vehicle_service_tokens', function (Blueprint $table) {
             $table->id();
             $table->integer('user_id');
-            $table->integer('vehicle_service_id')->unsigned();
+            $table->bigInteger('vehicle_service_id')->unsigned();
             $table->foreign('vehicle_service_id')->references('id')->on('vehicle_services')->onDelete('cascade');
             $table->string('email');
             $table->text('token');
